+++ py/desitarget/targets.py
@@ -25,7 +25,7 @@ log = get_logger()
 zcut = 2.1
 # ADM common redshift that defines a QSO to be reobserved for
 # ADM the Gontcho a Gontcho and Weiner et al. secondary programs.
-midzcut = 0.7
+midzcut = 1.6
 
 
 def encode_targetid(objid=None, brickid=None, release=None,
@@ -460,7 +460,7 @@ def calc_numobs_more(targets, zcat, obscon):
     obscon : :class:`str`
         A combination of strings that are in the desitarget bitmask yaml
         file (specifically in `desitarget.targetmask.obsconditions`), e.g.
-        "DARK|GRAY". Governs the behavior of how priorities are set based
+        "DARK". Governs the behavior of how priorities are set based
         on "obsconditions" in the desitarget bitmask yaml file.
 
     Returns
@@ -473,9 +473,9 @@ def calc_numobs_more(targets, zcat, obscon):
         - Will automatically detect if the passed targets are main
           survey, commissioning or SV and behave accordingly.
         - Most targets are updated to NUMOBS_MORE = NUMOBS_INIT-NUMOBS.
-          Special cases include BGS targets which always get NUMOBS_MORE
-          of 1 in bright time and "tracer" primary targets at z <
-          midzcut, which always get just one total observation.
+          Special cases for the main survey include BGS targets which
+          always get NUMOBS_MORE of 1 in bright time and "tracer" primary
+          targets at z < midzcut, which always get just one observation.
     """
     # ADM check input arrays are sorted to match row-by-row on TARGETID.
     assert np.all(targets["TARGETID"] == zcat["TARGETID"])
@@ -492,9 +492,10 @@ def calc_numobs_more(targets, zcat, obscon):
     # ADM main case, just decrement by NUMOBS.
     numobs_more = np.maximum(0, targets['NUMOBS_INIT'] - zcat['NUMOBS'])
 
-    if survey != 'cmx':
+    if survey != 'cmx' and survey != 'sv3':
         # ADM BGS targets are observed during the BRIGHT survey, regardless
         # ADM of how often they've previously been observed.
+        # ADM This behavios is turned off for SV3.
         if (obsconditions.mask(obscon) & obsconditions.mask("BRIGHT")) != 0:
             ii = targets[desi_target] & desi_mask.BGS_ANY > 0
             numobs_more[ii] = 1
@@ -585,8 +586,6 @@ def calc_priority(targets, zcat, obscon, state=False):
                             dtype=mtldatamodel["TARGET_STATE"].dtype)
 
     # Determine which targets have been observed.
-    # TODO: this doesn't distinguish between really unobserved vs not yet
-    # processed.
     unobs = (zcat["NUMOBS"] == 0)
     log.debug('calc_priority has %d unobserved targets' % (np.sum(unobs)))
     if np.all(unobs):
@@ -641,6 +640,11 @@ def calc_priority(targets, zcat, obscon, state=False):
                     ):
                         # ADM update priorities and target states.
                         Mxp = desi_mask[name].priorities[sname]
+                        # ADM tiered system in SV3. Decrement MORE_ZWARN
+                        # ADM priority using the bit's zwarndecrement.
+                        if survey == "sv3" and sname == "MORE_ZWARN":
+                            zwd = desi_mask[name].priorities["ZWARN_DECREMENT"]
+                            Mxp -= zwd * zcat[ii & sbool]["NUMOBS"]
                         # ADM update states BEFORE changing priorities.
                         ts = "{}|{}".format(name, sname)
                         target_state[ii & sbool] = np.where(
@@ -661,11 +665,6 @@ def calc_priority(targets, zcat, obscon, state=False):
                 good_midz = (zgood & (zcat['Z'] >= midzcut) &
                              (zcat['Z'] < zcut) & (zcat['ZWARN'] == 0))
 
-# ADM useful to turn on if, for example, we want to get lots
-# ADM of QSO repeat observations during the 1% Survey:
-#                if survey[:2] == 'sv':
-#                    good_hiz = zgood & (zcat['ZWARN'] == 0)
-
                 for sbool, sname in zip(
                         [unobs, done, good_hiz, good_midz,
                          ~good_hiz & ~good_midz, zwarn],
@@ -674,6 +673,11 @@ def calc_priority(targets, zcat, obscon, state=False):
                 ):
                     # ADM update priorities and target states.
                     Mxp = desi_mask[name].priorities[sname]
+                    # ADM tiered system in SV3. Decrement MORE_ZWARN
+                    # ADM priority using the bit's zwarndecrement.
+                    if survey == "sv3" and sname == "MORE_ZWARN":
+                        zwd = desi_mask[name].priorities["ZWARN_DECREMENT"]
+                        Mxp -= zwd * zcat[ii & sbool]["NUMOBS"]
                     # ADM update states BEFORE changing priorities.
                     ts = "{}|{}".format(name, sname)
                     target_state[ii & sbool] = np.where(
@@ -694,6 +698,11 @@ def calc_priority(targets, zcat, obscon, state=False):
                     ):
                         # ADM update priorities and target states.
                         Mxp = bgs_mask[name].priorities[sname]
+                        # ADM tiered system in SV3. Decrement MORE_ZWARN
+                        # ADM priority using the bit's zwarndecrement.
+                        if survey == "sv3" and sname == "MORE_ZWARN":
+                            zwd = bgs_mask[name].priorities["ZWARN_DECREMENT"]
+                            Mxp -= zwd * zcat[ii & sbool]["NUMOBS"]
                         # ADM update states BEFORE changing priorities.
                         ts = "{}|{}".format("BGS", sname)
                         target_state[ii & sbool] = np.where(
@@ -720,6 +729,11 @@ def calc_priority(targets, zcat, obscon, state=False):
                         ):
                             # ADM update priorities and target states.
                             Mxp = mws_mask[name].priorities[sname]
+                            # ADM tiered system in SV3. Decrement MORE_ZWARN
+                            # ADM priority using the bit's zwarndecrement.
+                            if survey == "sv3" and sname == "MORE_ZWARN":
+                                zwd = mws_mask[name].priorities["ZWARN_DECREMENT"]
+                                Mxp -= zwd * zcat[ii & sbool]["NUMOBS"]
                             # ADM update states BEFORE changing priorities.
                             ts = "{}|{}".format("MWS", sname)
                             target_state[ii & sbool] = np.where(
@@ -767,6 +781,11 @@ def calc_priority(targets, zcat, obscon, state=False):
                     for sbool, sname in zip(sbools, snames):
                         # ADM update priorities and target states.
                         Mxp = scnd_mask[name].priorities[sname]
+                        # ADM tiered system in SV3. Decrement MORE_ZWARN
+                        # ADM priority using the bit's zwarndecrement.
+                        # if survey == "sv3" and sname == "MORE_ZWARN":
+                        #    zwd = scnd_mask[name].priorities["ZWARN_DECREMENT"]
+                        #    Mxp -= zwd * zcat[ii & sbool]["NUMOBS"]
                         # ADM update states BEFORE changing priorities.
                         ts = "{}|{}".format("SCND", sname)
                         target_state[ii & sbool] = np.where(